# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API implementation.",
        "title": "Swagger System Payment",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/dlocal/card": {
            "post": {
                "description": "Creates card and saves card token in database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dlocal"
                ],
                "summary": "Create Card with Dlocal",
                "parameters": [
                    {
                        "description": "Card example",
                        "name": "example",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dlocal.Card"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dlocal.CardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError500"
                        }
                    }
                }
            }
        },
        "/dlocal/payment": {
            "post": {
                "description": "Makes a new Payment with the Dlocal API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dlocal"
                ],
                "summary": "Make Payment with Dlocal",
                "parameters": [
                    {
                        "description": "Payment with Dlocal example",
                        "name": "example",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dlocal.PaymentRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dlocal.PaymentResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError500"
                        }
                    }
                }
            }
        },
        "/dlocal/secure-payment": {
            "post": {
                "description": "Makes a new Secure Payment with the Dlocal API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dlocal"
                ],
                "summary": "Make Secure Payment with Dlocal",
                "parameters": [
                    {
                        "description": "Secure Payment with Dlocal example",
                        "name": "example",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dlocal.SecurePaymentRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dlocal.SecurePaymentRequestBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError500"
                        }
                    }
                }
            }
        },
        "/dummy/dummies": {
            "get": {
                "description": "Select all Dummies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dummy"
                ],
                "summary": "Select all Dummies",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 0,
                        "description": "start example",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "description": "count example",
                        "name": "count",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Dummy"
                            }
                        }
                    }
                }
            }
        },
        "/dummy/new": {
            "post": {
                "description": "save dummy in database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dummy"
                ],
                "summary": "Insert Dummy",
                "parameters": [
                    {
                        "description": "Dummy example",
                        "name": "example",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Dummy"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Dummy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError500"
                        }
                    }
                }
            }
        },
        "/dummy/update": {
            "put": {
                "description": "Updates a dummy in database (id req)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dummy"
                ],
                "summary": "Updates Dummy",
                "parameters": [
                    {
                        "description": "Dummy example",
                        "name": "example",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Dummy"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Dummy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError500"
                        }
                    }
                }
            }
        },
        "/dummy/{id}": {
            "get": {
                "description": "Get one Dummy from ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dummy"
                ],
                "summary": "Select Dummy",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "example: 1",
                        "name": "int",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Dummy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dlocal.Address": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "zipCode": {
                    "type": "string"
                }
            }
        },
        "dlocal.Card": {
            "type": "object",
            "properties": {
                "cvv": {
                    "type": "string",
                    "maxLength": 3,
                    "minLength": 3,
                    "example": "123"
                },
                "expiration_month": {
                    "type": "integer",
                    "maximum": 12,
                    "minimum": 1,
                    "example": 3
                },
                "expiration_year": {
                    "type": "integer",
                    "example": 2033
                },
                "holder_name": {
                    "type": "string",
                    "example": "Jhon Doe"
                },
                "number": {
                    "type": "string",
                    "example": "4111111111111111"
                }
            }
        },
        "dlocal.CardResponse": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "card_id": {
                    "type": "string"
                },
                "expiration_month": {
                    "type": "string"
                },
                "expiration_year": {
                    "type": "string"
                },
                "holder_name": {
                    "type": "string"
                },
                "last4": {
                    "type": "string"
                }
            }
        },
        "dlocal.Payer": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/dlocal.Address"
                },
                "birthDate": {
                    "type": "string"
                },
                "device_id": {
                    "type": "string"
                },
                "document": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 6
                },
                "ip": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 6
                },
                "phone": {
                    "type": "string"
                },
                "userReference": {
                    "type": "string"
                }
            }
        },
        "dlocal.PaymentRequestBody": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "card": {
                    "$ref": "#/definitions/dlocal.SecureCard"
                },
                "country": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "notification_url": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "payer": {
                    "$ref": "#/definitions/dlocal.Payer"
                },
                "payment_method_flow": {
                    "type": "string"
                },
                "payment_method_id": {
                    "type": "string"
                }
            }
        },
        "dlocal.PaymentResponseBody": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "approved_date": {
                    "type": "string"
                },
                "card": {
                    "$ref": "#/definitions/dlocal.CardResponse"
                },
                "country": {
                    "type": "string"
                },
                "created_date": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notification_url": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "payment_method_flow": {
                    "type": "string"
                },
                "payment_method_id": {
                    "type": "string"
                },
                "payment_method_type": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "status_code": {
                    "type": "string"
                },
                "status_detail": {
                    "type": "string"
                }
            }
        },
        "dlocal.SecureCard": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "dlocal.SecurePaymentRequestBody": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "card": {
                    "$ref": "#/definitions/dlocal.Card"
                },
                "country": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "notification_url": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "payer": {
                    "$ref": "#/definitions/dlocal.Payer"
                },
                "payment_method_flow": {
                    "type": "string"
                },
                "payment_method_id": {
                    "type": "string"
                }
            }
        },
        "httputil.HTTPError400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request body or query parameters"
                }
            }
        },
        "httputil.HTTPError404": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "error": {
                    "type": "string",
                    "example": "Page not found"
                }
            }
        },
        "httputil.HTTPError500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "error": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
...
        "model.Dummy": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "desktop chair"
                },
                "price": {
                    "type": "number",
                    "example": 299.99
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Description for what is this security definition being used",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": "\t\t\t\t\t\t\tGrants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": "\t\t\t\t\t\t\tGrants read and write access to administrative information",
                "write": "\t\t\t\t\t\t\tGrants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": "\t\t\t\t\t\t\tGrants read and write access to administrative information",
                "write": "\t\t\t\t\t\t\tGrants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
...
}